# packages/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alumni API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/me/": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Get Me",
        "operationId": "get_me_me__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      },
      "put": {
        "tags": [
          "me"
        ],
        "summary": "Update Me",
        "operationId": "update_me_me__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/me/directory-consent": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Submit Directory Consent",
        "operationId": "submit_directory_consent_me_directory_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectoryConsentWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/me/avatar": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Upload Avatar",
        "operationId": "upload_avatar_me_avatar_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_me_avatar_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/me/change-requests": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "List My Change Requests",
        "operationId": "list_my_change_requests_me_change_requests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileChangeRequestRead"
                  },
                  "type": "array",
                  "title": "Response List My Change Requests Me Change Requests Get"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      },
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Create Change Request",
        "operationId": "create_change_request_me_change_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileChangeRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileChangeRequestRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/me/view-requests": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "List Incoming View Requests",
        "operationId": "list_incoming_view_requests_me_view_requests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DirectoryViewRequestRead"
                  },
                  "type": "array",
                  "title": "Response List Incoming View Requests Me View Requests Get"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/me/view-requests/{request_id}/accept": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Accept View Request",
        "operationId": "accept_view_request_me_view_requests__request_id__accept_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryViewRequestRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/me/view-requests/{request_id}/decline": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Decline View Request",
        "operationId": "decline_view_request_me_view_requests__request_id__decline_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryViewRequestRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/me/view-requests/{request_id}/revoke": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Revoke View Request",
        "operationId": "revoke_view_request_me_view_requests__request_id__revoke_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryViewRequestRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_avatar_me_avatar_post": {
        "properties": {
          "avatar": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Avatar"
          }
        },
        "type": "object",
        "required": [
          "avatar"
        ],
        "title": "Body_upload_avatar_me_avatar_post"
      },
...
      "DirectoryConsentWrite": {
        "properties": {
          "visibility": {
            "type": "string",
            "enum": [
              "all",
              "cohort",
              "private"
            ],
            "title": "Visibility"
          }
        },
        "type": "object",
        "required": [
          "visibility"
        ],
        "title": "DirectoryConsentWrite"
      },
...
      "DirectoryViewRequestRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "requester_id": {
            "type": "integer",
            "title": "Requester Id"
          },
          "requester_name": {
            "type": "string",
            "title": "Requester Name"
          },
          "requester_cohort": {
            "type": "integer",
            "title": "Requester Cohort"
          },
          "target_id": {
            "type": "integer",
            "title": "Target Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "declined",
              "revoked"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "requester_id",
          "requester_name",
          "requester_cohort",
          "target_id",
          "status",
          "created_at"
        ],
        "title": "DirectoryViewRequestRead"
      },
...
      "MemberRead": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cohort": {
            "type": "integer",
            "title": "Cohort"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          },
          "roles": {
            "type": "string",
            "title": "Roles",
            "default": "member"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "all",
              "cohort",
              "private"
            ],
            "title": "Visibility",
            "default": "private"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "birth_lunar": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Lunar"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "company_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Phone"
          },
          "addr_personal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Addr Personal"
          },
          "addr_company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Addr Company"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "suspended",
              "rejected"
            ],
            "title": "Status",
            "default": "active"
          },
          "directory_consent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directory Consent At"
          },
          "directory_choice_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directory Choice At"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "email",
          "name",
          "cohort",
          "id",
          "avatar_url"
        ],
        "title": "MemberRead"
      },
...
      "MemberUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "all",
                  "cohort",
                  "private"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "birth_lunar": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Lunar"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "company_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Phone"
          },
          "addr_personal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Addr Personal"
          },
          "addr_company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Addr Company"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          }
        },
        "type": "object",
        "title": "MemberUpdate"
      },
...
      "ProfileChangeRequestCreate": {
        "properties": {
          "field_name": {
            "type": "string",
            "enum": [
              "name",
              "cohort"
            ],
            "title": "Field Name"
          },
          "new_value": {
            "type": "string",
            "maxLength": 255,
            "title": "New Value"
          }
        },
        "type": "object",
        "required": [
          "field_name",
          "new_value"
        ],
        "title": "ProfileChangeRequestCreate"
      },
...
      "ProfileChangeRequestRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "member_id": {
            "type": "integer",
            "title": "Member Id"
          },
          "field_name": {
            "type": "string",
            "enum": [
              "name",
              "cohort"
            ],
            "title": "Field Name"
          },
          "old_value": {
            "type": "string",
            "title": "Old Value"
          },
          "new_value": {
            "type": "string",
            "title": "New Value"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "title": "Status"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by_student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By Student Id"
          },
          "reject_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reject Reason"
          },
          "member_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Name"
          },
          "member_student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Student Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "member_id",
          "field_name",
          "old_value",
          "new_value",
          "status",
          "requested_at"
        ],
        "title": "ProfileChangeRequestRead"
      },
...
    },
    "responses": {
      "ProblemDetailsError": {
        "description": "Problem Details error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    }
  }
}